# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Minerva Credit Hunter API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: integer
        - name: assignedTo
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /product-categories:
    get:
      operationId: listProductCategories
      tags: [products]
      summary: List product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductCategory"
    post:
      operationId: createProductCategory
      tags: [products]
      summary: Create product category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductCategory"
...
components:
  schemas:
...
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, city, isActive, createdAt, updatedAt]
...
    User:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [superadmin, head_office_admin, editor, branch_head, hunter]
        branchId:
          type: integer
          nullable: true
        branch:
          $ref: "#/components/schemas/Branch"
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, telegramId, name, role, isActive, createdAt, updatedAt]
...
    Client:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        fullName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, questionnaire, recommendation, basket, pdf_generated, completed, rejected]
        branchId:
          type: integer
        branch:
          $ref: "#/components/schemas/Branch"
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        assignedTo:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sessionId, status, branchId, createdAt, updatedAt]
    ClientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [data, total, page, pageSize]
    CreateClientBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        branchId:
          type: integer
        assignedToId:
          type: integer
      required: [branchId]
    UpdateClientBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [draft, questionnaire, recommendation, basket, pdf_generated, completed, rejected]
        assignedToId:
          type: integer
    ProductCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    CreateProductCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [credit, non_credit]
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/ProductCategory"
          nullable: true
        description:
          type: string
          nullable: true
        minAmount:
          type: number
          nullable: true
        maxAmount:
          type: number
          nullable: true
        minTermMonths:
          type: integer
          nullable: true
        maxTermMonths:
          type: integer
          nullable: true
        interestRate:
          type: number
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, type, isActive, createdAt, updatedAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [credit, non_credit]
        categoryId:
          type: integer
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
        minTermMonths:
          type: integer
        maxTermMonths:
          type: integer
        interestRate:
          type: number
        isActive:
          type: boolean
      required: [name, type]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [credit, non_credit]
        categoryId:
          type: integer
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
        minTermMonths:
          type: integer
        maxTermMonths:
          type: integer
        interestRate:
          type: number
        isActive:
          type: boolean
...