# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minerva Credit Hunter API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProductBodyType } from "./createProductBodyType";

export interface CreateProductBody {
  name: string;
  type: CreateProductBodyType;
  categoryId?: number;
  description?: string;
  minAmount?: number;
  maxAmount?: number;
  minTermMonths?: number;
  maxTermMonths?: number;
  interestRate?: number;
  isActive?: boolean;
}