# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minerva Credit Hunter API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from "./productCategory";
import type { ProductType } from "./productType";

export interface Product {
  id: number;
  name: string;
  type: ProductType;
  categoryId?: number | null;
  category?: ProductCategory | null;
  description?: string | null;
  minAmount?: number | null;
  maxAmount?: number | null;
  minTermMonths?: number | null;
  maxTermMonths?: number | null;
  interestRate?: number | null;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}