# lib/api-zod/src/generated/types/updateClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minerva Credit Hunter API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateClientBodyStatus } from "./updateClientBodyStatus";

export interface UpdateClientBody {
  fullName?: string;
  phone?: string;
  status?: UpdateClientBodyStatus;
  assignedToId?: number;
}